requirements.txt
torch==2.5.1
torchvision==0.20.1
ultralytics==8.2.103
onnx==1.17.0
onnxruntime==1.20.1
opencv-python==4.10.0.84
numpy==1.26.4
python-dotenv==1.0.1
roboflow==1.1.44
